aws S3 수정

+++ backend/src/main/java/fullcare/backend/project/service/ProjectService.java
@@ -58,7 +58,7 @@ public class ProjectService {
         Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException("프로젝트 정보가 없습니다."));
         String imageUrl = project.getImageUrl();
         project.update(projectUpdateRequest);
-        if(imageUrl != null) {
+        if(imageUrl != null && !imageUrl.isEmpty()) {
             uploadService.delete(imageUrl);
         }
     }
@@ -71,7 +71,7 @@ public class ProjectService {
     public void deleteProject(Long projectId) {
         Project project = projectRepository.findById(projectId).orElseThrow(() -> new EntityNotFoundException("프로젝트 정보가 없습니다."));
         projectRepository.deleteById(projectId);
-        if (project.getImageUrl() != null) {
+        if (project.getImageUrl() != null && !project.getImageUrl().isEmpty()) {
             uploadService.delete(project.getImageUrl());
         }
     }
